+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-## dev-master (2017-10-10)
+## dev-master (2017-10-25)
 
-## 0.0.1 (2017-10-10)
+- Renamed the ArrayStorage class in StorageArray.
+- Added StorageList class.
+- Added unitTest for class Brownie\Util\StorageList.
+
+## dev-master (2017-10-10)
 - Initinal commit.
 - Added ArrayList class.
 - Added Exception\UndefinedMethodException class.